Enhance htmx_text_field function in widgets app
The improvements include the ability to accept arbitrary kwargs starting with 'hx-' which are used to customize the HTMX attributes. In addition, the app label is now included as part of the URL path using data extracted from the model object rather than being directly passed as an argument. This enhances the flexibility of URL generation and HTMX template usage.